Flattened Data
For the one file I took the average of the values in the composite
columns and used that as a single value for that row.

For the second file, I just completely flattened the data - each data
point in each composite column got its own row.

I did NOT do the time to end column yet because that is a separate
beast.

Note: I realize I could have done the first one with a for loop! I just
left it that way because I was halfway through and wanted to get it
done!!